Add Uncompress dicom function Rename the ImageProcess to ArrayProcess Rename the NiiImageProcess to ImageProcess
